refactor: rename 404 to ErrorPage
